+++ node/resolvers/catalog/index.ts
@@ -72,7 +72,7 @@ const translateToStoreDefaultLanguage = async (
     segment.getSegment(),
   ])
   return from && from !== to
-    ? messagesGraphQL.translate(toSearchTerm(term, from, to) as any).then(head)
+    ? messagesGraphQL.translateV2(toSearchTerm(term, from, to) as any).then(head)
     : term
 }