+++ node/utils/ioMessage.ts
@@ -1,62 +1,14 @@
-import { Segment } from '@vtex/api'
-import { prop } from 'ramda'
-import { Slugify } from './slug'
-
-export const localeFromDefaultSalesChannel = (segment: Segment) =>
-  segment.getSegmentByToken(null).then(prop('cultureInfo'))
-
-export const toIOMessage = async (
-  segment: Segment,
-  content: string,
-  id: string
-) => content
-
-export const toFacetIOMessage = (
-  segment: Segment,
-  content: string,
-  id: string
-) => toIOMessage(segment, content, `SpecificationFilter-id.${id}::${content}`)
-
-export const toClusterIOMessage = (
-  segment: Segment,
-  content: string,
-  id: string
-) => toIOMessage(segment, content, `ProductCluster-id.${id}::${content}`)
-
-export const toSKUIOMessage = (field: string) => (
-  segment: Segment,
-  content: string,
-  id: string
-) => toIOMessage(segment, content, `SKU-id.${id}::${field}`)
-
-export const toBrandIOMessage = (field: string) => (
-  segment: Segment,
-  content: string,
-  id: number | string
-) => toIOMessage(segment, content, `Brand-id.${id}::${field}`)
-
-export const toSpecificationIOMessage = (field: string) => (
-  segment: Segment,
-  content: string,
-  id: string
-) => toIOMessage(segment, content, `Specification-id.${id}::${field}`)
 
 export const toSearchTerm = (
   term: string,
   from: string,
   to: string,
-  description: string = ''
+  description?: string
 ) => ({
   messages: [
     {
-      provider: 'Search',
-      messages: [
-        {
-          id: Slugify(term),
-          content: term,
-          description,
-        },
-      ],
+      content: term,
+      context: description
     },
   ],
   from,